+++ samples/react-native/metro.config.js
@@ -2,10 +2,8 @@ const { getDefaultConfig, mergeConfig } = require('@react-native/metro-config');
 const path = require('path');
 const blacklist = require('metro-config/src/defaults/exclusionList');
 
-const {
-  createSentryMetroSerializer,
-} = require('../dist/js/tools/sentryMetroSerializer');
-const parentDir = path.resolve(__dirname, '..');
+const { createSentryMetroSerializer } = require('../../metro');
+const parentDir = path.resolve(__dirname, '../..');
 
 /**
  * Metro configuration